pyproject.toml
[tool.poetry]
name = "largebot"
version = "4.2.18"
description = ""
authors = ["welocalize <aie-devel@welocalize.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.7.2"
O365 = "^2.0.11"
selenium = "^3.141.0"
pathlib = "^1.0.1"
openpyxl = "^3.0.5"
typer = "^0.3.2"
pymsteams = "^0.1.14"
arrow = "^0.17.0"
num2words = "^0.5.10"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.0"
fastapi = "^0.63.0"
uvicorn = "^0.13.0"
gunicorn = "^20.0.4"
urllib3 = "^1.26.2"
tabulate = "^0.8.7"
aiofiles = "^0.6.0"
python-multipart = "^0.0.5"
"ruamel.yaml" = "^0.16.12"
rapidfuzz = "^0.14.1"
typer-cli = "^0.0.11"
PyYAML = "^5.3.1"
py2app = "^0.22"
numpy = "^1.19.4"
pandas = "^1.2.0"
wxPython = "^4.1.1"
PySimpleGUI = "^4.32.1"
Faker = "^5.0.2"
StringGenerator = "^0.3.4"
slot-faker = "^0.1.3"
jinja2 = "^2.11.2"
wolframalpha = "^4.1.1"
lxml = "^4.6.2"
inflect = "^5.0.2"
termcolor = "^1.1.0"
colorama = "^0.4.4"
faker-wifi-essid = "^0.2.0"
mypy = "^0.790"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# [[tool.poetry.source]]
# name = "default"
# url = "https://pypi.python.org/simple/"
# default = true
 
[[tool.poetry.source]]
name = "rypy"
url = "https://pypi.rypy.tk/simple/"
secondary = true

[tool.poetry.scripts]
largebot = "largebot.cli:cli"
